Cleaned up, formatted, and corrected some submissions.
