+++ QuestionSets/PythonLanguage/BasicsAndArrays/Manan_Garg/Answer2.py
@@ -1,5 +1,5 @@
 x = 0
 A = range(0, 1001)
 for i in A:
-    x += i
+	x += i
 print(x)
