+++ QuestionSets/PythonLanguage/BasicsAndArrays/Naman_Modani/Q2.py
@@ -1,22 +1,8 @@
-Python
-3.6
-.5(v3
-.6
-.5: f59c0932b4, Mar
-28
-2018, 17: 00:18) [MSC v.1900 64 bit(AMD64)]
-on
-win32
-Type
-"copyright", "credits" or "license()"
-for more information.
-    >> > n = 1000
->> > i = 1
->> > total = 0
+total = 0
+n = 1000
+i = 0
 while i <= n:
-    total += i
-    i += 1
+	total += i
+	i += 1
 
 print(total)
-500500
->> >
