+++ QuestionSets/PythonLanguage/BasicsAndArrays/Param_Kapur/Q2_Answer.py
@@ -2,7 +2,7 @@ n = 1000
 i = 1
 total = 0
 while i <= n:
-    total += i
-    i += 1
+	total += i
+	i += 1
 
 print(total)
