+++ QuestionSets/PythonLanguage/BasicsAndArrays/Param_Kapur/Q4_Answer.py
@@ -2,7 +2,7 @@ import numpy as np
 
 doo = np.array([], dtype=int)
 for i in np.arange(1, 11):
-    foo = np.array([np.arange(i, 11 * i, i)])
-    doo = np.append(doo, foo)
+	foo = np.array([np.arange(i, 11 * i, i)])
+	doo = np.append(doo, foo)
 doo = doo.reshape(10, 10)
 print(doo)
