+++ QuestionSets/PythonLanguage/BasicsAndArrays/Rajat_Varma/Answer_2.py
@@ -1,4 +1,4 @@
 count = 0
 for i in range(1, 1001):
-    count += i
+	count += i
 print(count)
